+++ htdocs/core/lib/date.lib.php
@@ -173,22 +173,22 @@ function dol_time_plus_duree($time, $duration_value, $duration_unit, $ruleforend
 	}
 	//Change the behavior of PHP over data-interval when the result of this function is Feb 29 (non-leap years), 30 or Feb 31 (so php returns March 1, 2 or 3 respectively)
 	if ($ruleforendofmonth == 1 && $duration_unit == 'm') {
-		$timeyear = dol_print_date($time, '%Y');
-		$timemonth = dol_print_date($time, '%m');
+		$timeyear = (int) dol_print_date($time, '%Y');
+		$timemonth = (int) dol_print_date($time, '%m');
 		$timetotalmonths = (($timeyear * 12) + $timemonth);
 
 		$monthsexpected = ($timetotalmonths + $duration_value);
 
 		$newtime = $date->getTimestamp();
 
-		$newtimeyear = dol_print_date($newtime, '%Y');
-		$newtimemonth = dol_print_date($newtime, '%m');
-		$newtimetotalmonths = (($newtimeyear * 12) + $newtimemonth);
+		$newtimeyear = (int) dol_print_date($newtime, '%Y');
+		$newtimemonth = (int) dol_print_date($newtime, '%m');
+		$newtimetotalmonths = (($newtimeyear * 12) +  $newtimemonth);
 
 		if ($monthsexpected < $newtimetotalmonths) {
-			$newtimehours = dol_print_date($newtime, '%H');
-			$newtimemins = dol_print_date($newtime, '%M');
-			$newtimesecs = dol_print_date($newtime, '%S');
+			$newtimehours = (int) dol_print_date($newtime, '%H');
+			$newtimemins = (int) dol_print_date($newtime, '%M');
+			$newtimesecs = (int) dol_print_date($newtime, '%S');
 
 			$datelim = dol_mktime($newtimehours, $newtimemins, $newtimesecs, $newtimemonth, 1, $newtimeyear);
 			$datelim -= (3600 * 24);
